+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added option `includeFiles` to only generate markdown for specified files
 
 ## [v0.4.1-6] - 2023-03-10
 - re-release due to technical problems in the release workflow (again)